+++ jsonargparse_tests/test_deprecated.py
@@ -57,6 +57,7 @@ from jsonargparse_tests.conftest import (
 )
 from jsonargparse_tests.test_dataclasses import DataClassA
 from jsonargparse_tests.test_jsonnet import example_2_jsonnet
+from jsonargparse_tests.test_paths import paths  # noqa: F401
 
 
 @pytest.fixture(autouse=True)
@@ -457,29 +458,29 @@ def test_ActionPath(tmp_cwd):
     parser.add_argument("--files", nargs="+", action=ActionPath(mode="fr"))
 
     cfg = parser.parse_args(["--cfg", abs_yaml_file])
-    assert str(tmp_cwd) == os.path.realpath(cfg.dir(absolute=True))
-    assert abs_yaml_file == os.path.realpath(cfg.cfg[0](absolute=False))
-    assert abs_yaml_file == os.path.realpath(cfg.cfg[0](absolute=True))
-    assert rel_yaml_file == cfg.file(absolute=False)
-    assert abs_yaml_file == os.path.realpath(cfg.file(absolute=True))
+    assert str(tmp_cwd) == os.path.realpath(cfg.dir.absolute)
+    assert abs_yaml_file == os.path.realpath(cfg.cfg[0].relative)
+    assert abs_yaml_file == os.path.realpath(cfg.cfg[0].absolute)
+    assert rel_yaml_file == cfg.file.relative
+    assert abs_yaml_file == os.path.realpath(cfg.file.absolute)
     pytest.raises(ArgumentError, lambda: parser.parse_args(["--cfg", abs_yaml_file + "~"]))
 
     cfg = parser.parse_args(["--cfg", "file: " + abs_yaml_file + "\ndir: " + str(tmp_cwd) + "\n"])
-    assert str(tmp_cwd) == os.path.realpath(cfg.dir(absolute=True))
+    assert str(tmp_cwd) == os.path.realpath(cfg.dir.absolute)
     assert cfg.cfg[0] is None
-    assert abs_yaml_file == os.path.realpath(cfg.file(absolute=True))
+    assert abs_yaml_file == os.path.realpath(cfg.file.absolute)
     pytest.raises(ArgumentError, lambda: parser.parse_args(["--cfg", '{"k":"v"}']))
 
     cfg = parser.parse_args(["--file", abs_yaml_file, "--dir", str(tmp_cwd)])
-    assert str(tmp_cwd) == os.path.realpath(cfg.dir(absolute=True))
-    assert abs_yaml_file == os.path.realpath(cfg.file(absolute=True))
+    assert str(tmp_cwd) == os.path.realpath(cfg.dir.absolute)
+    assert abs_yaml_file == os.path.realpath(cfg.file.absolute)
     pytest.raises(ArgumentError, lambda: parser.parse_args(["--dir", abs_yaml_file]))
     pytest.raises(ArgumentError, lambda: parser.parse_args(["--file", str(tmp_cwd)]))
 
     cfg = parser.parse_args(["--files", abs_yaml_file, abs_yaml_file])
     assert isinstance(cfg.files, list)
     assert 2 == len(cfg.files)
-    assert abs_yaml_file == os.path.realpath(cfg.files[-1](absolute=True))
+    assert abs_yaml_file == os.path.realpath(cfg.files[-1].absolute)
 
     pytest.raises(TypeError, lambda: parser.add_argument("--op1", action=ActionPath))
     pytest.raises(
@@ -543,6 +544,19 @@ def test_Path_attr_set(tmp_cwd):
         assert path.abs_path == os.path.join(tmp_cwd, "file")
         assert path.skip_check is False
         assert "Path objects are not meant to be mutable" in str(w[-1].message)
+
+
+def test_path_call(paths):  # noqa: F811
+    path = Path(paths.file_rw, "frw")
+    with catch_warnings(record=True) as w:
+        assert path(False) == str(paths.file_rw)
+    assert_deprecation_warn(
+        w,
+        message="Calling Path objects is deprecated",
+        code="assert path(False) == ",
+    )
+    assert path(True) == str(paths.tmp_path / paths.file_rw)
+    assert path() == str(paths.tmp_path / paths.file_rw)
 
 
 def test_ActionPathList(tmp_cwd):